fix temp file remove error

+++ core/utils/database.py
@@ -94,10 +94,11 @@ def bulk_insert_from_file(model, file_path, **kwargs):
             rows = model.transform_self_from_file(file_path, kwargs['update'])
             batch_upsert_from_gen(model, rows, settings.BATCH_SIZE, **kwargs)
 
-    os.remove(temp_file_path)
-
     if 'callback' in kwargs and kwargs['callback']:
         kwargs['callback']()
+
+    if os.path.isfile(temp_file_path):
+        os.remove(temp_file_path)
 
 
 def copy_insert_from_csv(table_name, temp_file_path, **kwargs):
